# Opportunistically decode attributes in AttributeDecoder