@ti-mo 2019-12-06
> LGTM overall, just a couple of nits. I assume your fuzzing is going well?

At about an hour now, had to interrupt but resumed from the same corpus.

`2019/12/06 16:00:16 workers: 16, corpus: 76 (2m46s ago), crashers: 0, restarts: 1/9994, execs: 301119156 (149143/sec), cover: 87, uptime: 33m39s`

Thanks for the review!