@ti-mo 2019-12-06
Fuzzer ran for 4.5hrs without any crashers. :point_up: