@ti-mo 2019-12-07
Looks like this implements https://github.com/mdlayher/netlink/issues/115!